Improved player layout & design; Server can now send custom buttons over each player to the client
The leftmost button is reserved for Acceptance/Refusal/Rent, which was formerly handled by the client's multibutton. Due to complications when multiple players used a Just Say No card against a rent charge, I've decided the best way to solve the issue was to apply buttons over every player. There's 3 buttons over every player in total, so the 2 rightmost buttons can be used for custom actions in mods.