Added first run screen, allowing users to select a place to store game data
Users can select between storing in the application data directory or in the same folder as where the game is being run from. On launch, the game will first check if data is present in its current directory, then if not present, check the application data directory. If neither are present, the first run screen will show.

Other notable changes:
- Font heights display more accurately across different systems(wish there was a better way to do it)
- Window is better centered
- Players are notified when they've been kicked for logging in from another client
- Players disconnected for invalid packets aren't removed from the game