Significant optimizations to the client
A massive problem is that components behind a moving card have to redraw entirely, causing horrific resource usage. I don't think there's a way to make the performance *good*, but I'm now caching every component's graphics so that at least the painting code isn't called a million times. Unfortunately, Swing is not meant for games and migration to a better API is required for more acceptable performance.